refactor precommit and transfer into cli
