+++ client/drs-map.go
@@ -107,6 +107,15 @@ func UpdateDrsMap() error {
 		}
 		fileURL := fmt.Sprintf("s3://%s", filepath.Join(bucketName, drsId, file.Oid))
 
+		// create authz string from profile
+		fmt.Println("cfg.Gen3Project:", cfg.Gen3Project)
+		// check if project ID is valid
+		if !strings.Contains(cfg.Gen3Project, "-") {
+			return fmt.Errorf("error: invalid project ID %s in config file, ID should look like <program>-<project>", cfg.Gen3Project)
+		}
+		projectIdArr := strings.SplitN(cfg.Gen3Project, "-", 2)
+		authzStr := "/programs/" + projectIdArr[0] + "/projects/" + projectIdArr[1]
+
 		// If the oid exists in drsMap, check if it matches the calculated uuid
 		// TODO: naive method, where only the first file with the same oid is stored
 		// in the future, will need to handle multiple files with the same oid
@@ -122,7 +131,7 @@ func UpdateDrsMap() error {
 				URLs:     []string{fileURL},
 				Hashes:   HashInfo{SHA256: file.Oid},
 				Size:     file.Size,
-				Authz:    []string{repoName},
+				Authz:    []string{authzStr},
 				// ContentCreatedDate: modDate,
 				// ContentUpdatedDate: modDate,
 			}
