+++ client/transfer.log
@@ -1,0 +1,24 @@
+2025/06/17 17:29:06 updateDrsMap started
+2025/06/17 17:29:06 git lfs ls-files output: {
+ "files": null
+}
+
+2025/06/17 17:29:06 Repo Name: git-gen3
+2025/06/17 17:29:06 Writing drsMap to /var/folders/nq/88_4pk_s25z4g3g52gvm5b88px118k/T/drs-map-4230269377.json
+2025/06/17 17:29:06 Updated drs-map.json with 0 entries
+2025/06/17 17:30:13 updateDrsMap started
+2025/06/17 17:30:13 git lfs ls-files output: {
+ "files": null
+}
+
+2025/06/17 17:30:13 Repo Name: git-gen3
+2025/06/17 17:30:13 Writing drsMap to /var/folders/nq/88_4pk_s25z4g3g52gvm5b88px118k/T/drs-map-3449762568.json
+2025/06/17 17:30:13 Updated drs-map.json with 0 entries
+2025/06/17 17:32:20 updateDrsMap started
+2025/06/17 17:32:20 git lfs ls-files output: {
+ "files": null
+}
+
+2025/06/17 17:32:20 Repo Name: git-gen3
+2025/06/17 17:32:20 Writing drsMap to /var/folders/nq/88_4pk_s25z4g3g52gvm5b88px118k/T/drs-map-1558809153.json
+2025/06/17 17:32:20 Updated drs-map.json with 0 entries
